Wow - lots of things:
Makefile - added libopenmpi-dev.
fifo: ==None to is not None
mpiWrapper - environ for no MPI.
readconfig - ==None to is not None
interpmodule - working on hex DM.
interpolate - hex DM
infScrn ==None to is not None
iscrn - ==None to is not None, added default scrnDir.
atmos - int wrapper around numpy.floor for phsShiftX/Y.
dm - ==None to is not None, hex DM.
elong - added sig to line profile
tel - added height and fov and outside value for pupil, also multi-pupil.

+++ util/atmos.py
@@ -700,7 +700,7 @@ class atmos:
                         interpPosCol+=interpPosColDict[key]
                     else:
                         interpPosCol+=1
-                phsShiftY=numpy.floor(interpPosCol)
+                phsShiftY=int(numpy.floor(interpPosCol))
                 tmpcol=interpPosCol
                 interpPosCol-=phsShiftY
 
@@ -712,7 +712,7 @@ class atmos:
                         interpPosRow+=interpPosRowDict[key]
                     else:
                         interpPosRow+=1
-                phsShiftX=numpy.floor(interpPosRow)
+                phsShiftX=int(numpy.floor(interpPosRow))
                 tmprow=interpPosRow
                 interpPosRow-=phsShiftX
